feat(service3d): new feat
