+++ index.html
@@ -15,7 +15,7 @@
 	<link rel="mask-icon" href="./images/favicon/safari-pinned-tab.svg" color="#5bbad5">
 	<meta name="msapplication-TileColor" content="#b91d47">
 	<meta name="theme-color" content="#ffffff">
-	<script src="script/main.js" defer></script>
+	<script src="dist/main.js" defer></script>
 </head>
 
 <body>
