+++ package.json
@@ -1,0 +1,23 @@
+{
+  "name": "services_3d_model",
+  "version": "1.0.0",
+  "description": "3dGlo",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/andrewBar-ko/services3d.git"
+  },
+  "author": "andrew barkovets",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/andrewBar-ko/services3d/issues"
+  },
+  "homepage": "https://github.com/andrewBar-ko/services3d#readme",
+  "devDependencies": {
+    "webpack": "^5.5.0",
+    "webpack-cli": "^4.2.0"
+  }
+}
