+++ src/index.js
@@ -1,0 +1,23 @@
+
+
+import countTimer from './modules/countTimer';
+import toggleMenu from './modules/toggleMenu';
+import togglePopUp from './modules/togglePopUp';
+import tabs from './modules/tabs';
+import addDot from './modules/addDot';
+import setCommandImg from './modules/setCommandImg';
+import calc from './modules/calc';
+import sendForm from './modules/sendForm';
+import slider from './modules/slider';
+import preoladerForm from './modules/preolader';
+
+countTimer('31 december 2020');
+toggleMenu();
+togglePopUp();
+tabs();
+addDot();
+setCommandImg();
+calc(100);
+sendForm();
+slider();
+preoladerForm();
