+++ src/modules/calc.js
@@ -1,0 +1,77 @@
+const calc = (price = 100) => {
+    const calcBlock = document.querySelector('.calc-block'),
+        calcType = document.querySelector('.calc-type'),
+        calcSquare = document.querySelector('.calc-square'),
+        calcDay = document.querySelector('.calc-day'),
+        calcCount = document.querySelector('.calc-count'),
+        totalValue = document.getElementById('total');
+
+    const countSum = () => {
+        let total = 0,
+            countValue = 1,
+            dayValue = 10;
+        const typeValue = calcType.options[calcType.selectedIndex].value,
+            squareValue = +calcSquare.value;
+
+
+        if (calcCount.value > 1) {
+            countValue += (calcCount.value - 1) / 10;
+        }
+
+        if (calcDay.value && calcDay.value < 5) {
+            dayValue *= 2;
+        }
+        if (calcDay.value && calcDay.value < 10) {
+            dayValue *= 1.5;
+        }
+
+        if (!!typeValue && !!squareValue) {
+            total = price * typeValue * squareValue * countValue * dayValue;
+        }
+
+        const animeTotal = (elem, value) => {
+            let push = value / 100;
+
+            const interval = setInterval(() => {
+
+                if (+elem.textContent >= value) {
+                    elem.textContent = value;
+                    clearInterval(interval);
+                } else {
+                    elem.textContent = Math.round(+elem.textContent + push);
+                    push += elem.textContent / 100;
+                }
+
+            }, 10);
+
+            return Math.round(elem.textContent);
+        };
+
+        totalValue.textContent = animeTotal(totalValue, total);
+    };
+
+    calcBlock.addEventListener('input', e => {
+        const target = e.target;
+
+        if (target.matches('select') ||
+        target.matches('input')) {
+            countSum();
+        }
+    });
+    // Enter Only Numbers!
+    const enterOnlyNumbers = () => {
+
+        calcBlock.addEventListener('input', e => {
+
+            if (e.target.matches('.calc-square') ||
+            e.target.matches('.calc-count') ||
+            e.target.matches('.calc-day')) {
+                e.target.value = e.target.value.replace(/\D/g, '');
+            }
+        });
+    };
+    enterOnlyNumbers();
+
+};
+
+export default calc;
