+++ src/modules/preolader.js
@@ -1,0 +1,12 @@
+const preoladerForm = () => {
+    const preloader = document.querySelector('#page-preloader'),
+        spinner = document.querySelector('.cssload-box-loading');
+
+    setTimeout(() => {
+        preloader.style.display = 'none';
+        spinner.style.display = 'none';
+    }, 500);
+
+};
+
+export default preoladerForm;
